+++ lib/rails_audit_log/graphql.rb
@@ -1,10 +1,12 @@
 # frozen_string_literal: true
 
+require "graphql"
 require_relative "graphql/version"
+require_relative "graphql/types/base_object"
+require_relative "graphql/types/audit_log_entry_type"
 
 module RailsAuditLog
   module Graphql
     class Error < StandardError; end
-    # Your code goes here...
   end
 end